refactor: share callback and wrapper
